peripheral: pre/post actions moved to register
- pre-read & post_write actions moved to register,
inside read/write

- access flags as 64-bits
- prev_value added to register, to compute changes
- register code restructured

+++ include/hw/misc/peripheral.h
@@ -86,7 +86,7 @@ typedef struct {
     MemoryRegion mmio;
 
     /* Default access flags, when registers do not define them. */
-    uint32_t default_access_flags;
+    uint64_t default_access_flags;
 
     /* All peripheral registers have the same size. (1-8) */
     uint32_t register_size_bytes;
